# docs/swagger/swagger.yaml
...
openapi: 3.0.1
info:
  title: FootballCheck
  version: '1.0'
  description: The API for the FootballCheck project
paths:
...
  "/api/v1/squads":
    get:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: sortState
        in: query
        schema:
          "$ref": "#/components/schemas/SquadSortState"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/SquadDto"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/SquadDto"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/SquadDto"
        '401':
          description: Unauthorized
    post:
      security:
      - Bearer: []
      tags:
      - Squad
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SquadBaseDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/SquadBaseDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/SquadBaseDto"
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/squads/{id}":
    patch:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SquadBaseDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/SquadBaseDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/SquadBaseDto"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
  "/api/v1/squads/{squadId}/players":
    get:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: squadId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: ClubName
        in: query
        schema:
          type: string
      - name: Surname
        in: query
        schema:
          type: string
      - name: Country
        in: query
        schema:
          type: string
      - name: MinPrice
        in: query
        schema:
          type: integer
          format: int32
      - name: MaxPrice
        in: query
        schema:
          type: integer
          format: int32
      - name: MinRating
        in: query
        schema:
          type: integer
          format: int32
      - name: MaxRating
        in: query
        schema:
          type: integer
          format: int32
      - name: sortState
        in: query
        schema:
          "$ref": "#/components/schemas/PlayerSortState"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/PlayerDto"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/PlayerDto"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/PlayerDto"
        '401':
          description: Unauthorized
    post:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: squadId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PlayerIdDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/PlayerIdDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/PlayerIdDto"
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/squads/{squadId}/coach":
    get:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: squadId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/CoachDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/CoachDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/CoachDto"
        '401':
          description: Unauthorized
    post:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: squadId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CoachIdDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/CoachIdDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CoachIdDto"
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/squads/{squadId}/players/{playerId}":
    delete:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: squadId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: playerId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/squads/{squadId}/coach/{coachId}":
    delete:
      security:
      - Bearer: []
      tags:
      - Squad
      parameters:
      - name: squadId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: coachId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/SquadDto"
        '401':
          description: Unauthorized
        '409':
          description: Conflict
...
components:
  schemas:
...
    CoachDto:
      type: object
      properties:
        surname:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        workExperience:
          type: integer
          format: int32
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
    CoachIdDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
      additionalProperties: false
...
    PlayerDto:
      type: object
      properties:
        clubId:
          type: integer
          format: int32
          nullable: true
        surname:
          type: string
          nullable: true
        rating:
          type: integer
          format: int32
          nullable: true
        country:
          type: string
          nullable: true
        price:
          type: integer
          format: int32
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
    PlayerIdDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
      additionalProperties: false
    PlayerSortState:
      enum:
      - IdAsc
      - IdDesc
      - SurnameAsc
      - SurnameDesc
      - RatingAsc
      - RatingDesc
      - CountryAsc
      - CountryDesc
      - ClubNameAsc
      - ClubNameDesc
      - PriceAsc
      - PriceDesc
      type: string
    SquadBaseDto:
      type: object
      properties:
        coachId:
          type: integer
          format: int32
          nullable: true
        name:
          type: string
          nullable: true
        rating:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    SquadDto:
      type: object
      properties:
        coachId:
          type: integer
          format: int32
          nullable: true
        name:
          type: string
          nullable: true
        rating:
          type: integer
          format: int32
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
    SquadSortState:
      enum:
      - IdAsc
      - IdDesc
      - CoachSurnameAsc
      - CoachSurnameDesc
      - NameAsc
      - NameDesc
      - RatingAsc
      - RatingDesc
      type: string
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: Please insert JWT with Bearer into field
      name: Authorization
      in: header